Add email existence check method to LoginController and tests
A method checkEmail has been added to LoginController that checks if a provided email exists in the system. Corresponding test cases have been added to LoginControllerTest to ensure the new endpoint responds correctly. This method will be used to validate whether or not an email has been registered in the system.