Add email existence check to UserService and tests
A method isEmailExist has been added to UserService, for checking if an email already exists in system. This change includes corresponding changes in 'UserServiceImpl' to link with the userDao. Additionally, unit tests have been added to 'UserServiceTest' to validate the new method. This will enhance email validation during user registration.