Add VirtuGatherEvent model
A new Java class, VirtuGatherEvent, has been added in the model package. This class represents the properties and behaviors of a virtual gathering event, including fields like title, start time, end time, location, and others. The added class will be used for managing and storing event data in the application.